Introduce ko_build, deprecate ko_image
