+++ docs/resources/build.md
@@ -1,22 +1,16 @@
 ---
 # generated by https://github.com/hashicorp/terraform-plugin-docs
-page_title: "ko_image Resource - terraform-provider-ko"
+page_title: "ko_build Resource - terraform-provider-ko"
 subcategory: ""
 description: |-
-  Deprecated: use ko_build
+  Sample resource in the Terraform provider scaffolding.
 ---
 
-# ko_image (Resource)
+# ko_build (Resource)
 
-Deprecated: use ko_build
+Sample resource in the Terraform provider scaffolding.
 
-## Example Usage
 
-```terraform
-resource "ko_image" "example" {
-  importpath = "github.com/google/ko"
-}
-```
 
 <!-- schema generated by tfplugindocs -->
 ## Schema
