+++ internal/provider/provider.go
@@ -54,6 +54,7 @@ func New(version string) func() *schema.Provider {
 			},
 			ResourcesMap: map[string]*schema.Resource{
 				"ko_image":   resourceImage(),
+				"ko_build":   resourceBuild(),
 				"ko_resolve": resolveConfig(),
 			},
 		}
